+++ Cargo.toml
@@ -1,29 +1,3 @@
-[package]
-name = "c64emu"
-description = "DDR's Commodore 64 emulator"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-clap = { version = "4.4.2", features = ["derive"] }
-colored = "2"
-cursive = { version="0.20", default-features=false, features=["termion-backend"] }
-cursive_hexview = { git="https://github.com/ddrcode/cursive_hexview.git", branch="start-addr" }
-# cursive_hexview = "0.7"
-substring = "1.4.5"
-log = "0.4.20"
-env_logger = "0.10.0"
-
-[[bin]]
-name = "machine"
-path = "src/main.rs"
-
-[[bin]]
-name = "c64"
-path = "src/main-c64.rs"
-
-[[bin]]
-name = "gui"
-path = "src/main-gui.rs"
+[workspace]
+members = ["crates/mos6502", "crates/c64", "crates/gui"]
+resolver = "2"
