+++ crates/c64/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "c64"
+authors = ["ddrcode"]
+description = "DDR's Commodore 64 emulator"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+machine = { path="../mos6502/" }
+clap = { version = "4.4.2", features = ["derive"] }
+colored = "2"
+substring = "1.4.5"
+env_logger = "0.10.0"
+
+[lib]
+name = "c64"
+path = "src/lib.rs"
+
+[[bin]]
+name = "c64"
+path = "src/main.rs"
+
