+++ crates/c64/src/lib.rs
@@ -1,0 +1,2 @@
+mod c64;
+pub use c64::*;
