+++ crates/gui/src/gui/cpu_state.rs
