+++ crates/gui/src/gui/memory_view.rs
