+++ crates/mos6502/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "machine"
+authors = ["ddrcode"]
+description = "DDR's Commodore 64 emulator"
+version = "0.1.0"
+edition = "2021"
+
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+clap = { version = "4.4.2", features = ["derive"] }
+colored = "2"
+substring = "1.4.5"
+env_logger = "0.10.0"
+
+[[bin]]
+name = "machine"
+path = "src/main.rs"
+
+[lib]
+name = "machine"
+path = "src/lib.rs"
+
+
