+++ crates/mos6502/src/cli/cli_utils.rs
