+++ crates/mos6502/src/lib.rs
@@ -1,0 +1,7 @@
+mod machine;
+pub mod utils;
+pub mod cli;
+pub mod mos6510;
+
+pub use machine::*;
+
