+++ crates/mos6502/src/machine.rs
