+++ crates/mos6502/src/machine/machine.rs
