+++ crates/mos6502/src/machine/machine_events.rs
