+++ crates/mos6502/src/machine/macros.rs
