+++ crates/mos6502/src/machine/memory.rs
