+++ crates/mos6502/src/main.rs
@@ -1,12 +1,10 @@
 extern crate colored;
 
-mod cli_args;
-mod cli_utils;
+mod cli;
 mod machine;
 mod mos6510;
 
-use crate::cli_args::Args;
-use crate::cli_utils::get_file_as_byte_vec;
+use crate::cli::{ Args, get_file_as_byte_vec };
 use crate::machine::{MOS6502Machine, Machine, MachineConfig};
 use clap::Parser;
