+++ crates/mos6502/src/utils.rs
